CAI/requirements.txt
# Core ML/NLP Libraries
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Vector Database & Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
chromadb>=0.4.0

# Data Collection
wikipedia-api>=0.6.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Text Processing
nltk>=3.8.0
tiktoken>=0.5.0

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Visualization & UI
streamlit>=1.28.0
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.13.0

# Report Generation
fpdf2>=2.7.0
jinja2>=3.1.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
